async_backend/requirements.txt
flask>=3.0
google-cloud-firestore>=2.16
google-genai>=1.0
python-dotenv>=1.0
